src/app/shared/services/unique-id/unique-id.service.spec.ts
import { UniqueIdService } from "./unique-id.service";

describe(UniqueIdService.name, () => {
  let service: UniqueIdService;

  beforeEach(() => {
    service = new UniqueIdService();
  });

  it(`#${UniqueIdService.prototype.generateUniqueIdWithPrefix.name}
  should generate ID when called with prefix`, () => {
    const id = service.generateUniqueIdWithPrefix("ng");
    expect(id.startsWith("ng-")).toBeTrue();
  });

  it(`#${UniqueIdService.prototype.generateUniqueIdWithPrefix.name}
  should not generate duplicated IDs when called multiple times `, () => {
    const ids = new Set();
    for (let i = 0; i < 50; i++) {
      ids.add(service.generateUniqueIdWithPrefix("ng"));
    }
    expect(ids.size).toBe(50);
  });

  it(`#${UniqueIdService.prototype.getNumberOfGeneratedUniqueIds.name}
  should return the number of generated IDs when called`, () => {
    service.generateUniqueIdWithPrefix("ng");
    service.generateUniqueIdWithPrefix("ng");
    expect(service.getNumberOfGeneratedUniqueIds()).toBe(2);
  });

  it(`#${UniqueIdService.prototype.generateUniqueIdWithPrefix.name}
  should throw something when called with empty values`, () => {
    const emptyValues = [ null, undefined, "", "0", "1" ];
    emptyValues.forEach(emptyValue => {
      expect(() => service.generateUniqueIdWithPrefix(emptyValue))
      .withContext(`Empty value: ${emptyValue}`)
      .toThrow();
    });
  });

});
